PDFBOX-815: PDFont.getEncodingManager is not thread safe
Use singleton instances for the EncodingManager and default encoding classes.
Replace the static encodings map in EncodingManager with if clauses since we only need four entries.
Remove the "protected static" getEncodingManager() method from PDFont. This was introduced in PDFBOX-601 and the remove should not cause backwards compatibility problems even though strictly speaking it is a change in the binary API.

git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@992271 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: a1a44d3824a6c8eb4e3f49a41c751613df096619